multiple fixes/additions while implementing #pqc-tls #injection

+++ src/main/java/org/bouncycastle/tls/SignatureScheme.java
@@ -200,7 +200,7 @@ public class SignatureScheme
     public static short getHashAlgorithm(int signatureScheme)
     {
         if (InjectedSigAlgorithms.isSigSchemeSupported(signatureScheme))
-            return InjectedSigAlgorithms.getCryptoHashAlgorithmIndex(signatureScheme); // #pqc-tls #injection
+            return (short)InjectedSigAlgorithms.getCryptoHashAlgorithmIndex(signatureScheme); // #pqc-tls #injection
         return (short)((signatureScheme >>> 8) & 0xFF);
     }
